requirements.txt
pyyaml
requests
tqdm
pick
prettytable
windows-curses
splunk-sdk
pycryptodome